requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
cryptography==41.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1
WTForms==3.1.1
gunicorn==21.2.0
